+++ template-parts/download-links/error.php
@@ -4,9 +4,15 @@
  *
  * @package Webbooks
  *
- * @var array<string, mixed> $args
+ * @var array<string, mixed> $args Template arguments.
  */
 
 $message = (string) ( $args['message'] ?? '' );
 ?>
-<div class="container-fluid mrg-tb"><div class="row"><div class="alert alert-danger" role="alert"><?php echo esc_html( $message ); ?></div></div></div>
+<div class="container-fluid mrg-tb">
+	<div class="row">
+		<div class="alert alert-danger" role="alert">
+			<?php echo esc_html( $message ); ?>
+		</div>
+	</div>
+</div>