+++ template-parts/layout/portfolio-footer.php
@@ -10,7 +10,15 @@
 		<!-- Footer -->
 		<div id="footer">
 			<ul class="copyright">
-				<li>&copy; Optimtzed <a href="https://www.facebook.com/homeandriy">Andriy Beznosko</a>. All rights reserved.</li><li>Design: <a href="http://html5up.net">HTML5 UP</a></li>
+				<li>
+					&copy; Optimtzed
+					<a href="https://www.facebook.com/homeandriy">Andriy Beznosko</a>.
+					All rights reserved.
+				</li>
+				<li>
+					Design:
+					<a href="http://html5up.net">HTML5 UP</a>
+				</li>
 			</ul>
 		</div>
 		<?php wp_footer(); ?>